+++ src/components/FavoriteCards/FavoriteCards.module.scss
@@ -42,7 +42,7 @@
   }
 
   &__hide {
-    display: none !important;
+    display: none;
   }
 }
