+++ src/pages/Filters/Filters.module.scss
@@ -1,4 +1,4 @@
-@import "../style-dev/vars";
+@import "../../style-dev/vars";
 
 .card__wrapper {
   position: relative;
