+++ src/pages/Game/Game.module.scss
