+++ src/pages/Game/Game.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import s from './Game.module.scss';
-import Picture from '../components/Picture/Picture';
-import { TreeContext } from '../context';
-import Settings from '../components/Settings/Settings';
-import Header from '../components/Header/Header';
-import FavoriteCards from '../components/FavoriteCards/FavoriteCards';
-import data from '../assets/data';
-import { IToyOnTree } from '../types/ITypes';
+import Picture from '../../components/Picture/Picture';
+import { TreeContext } from '../../context';
+import Settings from '../../components/Settings/Settings';
+import Header from '../../components/Header/Header';
+import FavoriteCards from '../../components/FavoriteCards/FavoriteCards';
+import data from '../../assets/data';
+import { IToyOnTree } from '../../types/ITypes';
 
 const Game = () => {
   const [isSnow, setIsSnow] = useState<boolean>(false);
