+++ src/pages/Home/Home.module.scss
@@ -1,4 +1,4 @@
-@import "../style-dev/vars";
+@import "../../style-dev/vars";
 
 .home {
   display: flex;
@@ -28,7 +28,7 @@
     top: 60px;
     width: 75px;
     height: 75px;
-    background-image: url("../assets/home/christmas-tree.png");
+    background-image: url("../../assets/home/christmas-tree.png");
     background-position: center center;
     background-size: cover;
